phc_microservice_ancservice/src/main/java/com/ssf/ancservice/config/AuditAwareImpl.java
package com.ssf.ancservice.config;

import com.ssf.ancservice.constant.ANCServiceConst;
import com.ssf.ancservice.utils.Utils;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.data.domain.AuditorAware;

import javax.servlet.http.HttpServletRequest;
import java.util.Optional;

public class AuditAwareImpl implements AuditorAware<String> {

    @Autowired
    HttpServletRequest request;

    @Override
    public Optional<String> getCurrentAuditor() {
        String idToken = request.getHeader(ANCServiceConst.XUSER_ID_HEADER);
        return Optional.of(Utils.getUserIdFromIdToken(idToken));
    }
}
